fairchem_local_server/requirements-test.txt
fastapi
uvicorn
pytest
httpx
ray[serve]
ase
torch
fairchem-core
